tests: Move CorruptBBolt to testutil
